requirements.txt
boto3>=1.26.0
openai>=1.0.0
python-dotenv>=0.19.0
pytest>=7.0.0